+++ src/cobib/utils/logging.py
@@ -1,15 +1,17 @@
 """coBib's logging module.
 
-This module provides utility methods to set up logging to different handlers.
+This module provides utility methods to set up logging to different handlers as well as a custom
+logging handler for prettified log formatting.
 """
 
 from __future__ import annotations
 
 import logging
 import logging.handlers
-import sys
 from importlib import metadata
+from io import StringIO
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from rich.console import ConsoleRenderable, Group
 from rich.markdown import Markdown
@@ -17,7 +19,11 @@ from rich.panel import Panel
 from rich.text import Text
 from typing_extensions import override
 
+from .console import PromptConsole
 from .rel_path import RelPath
+
+if TYPE_CHECKING:
+    from cobib.ui.ui import UI
 
 HINT = 35
 """The logging level value for HINT messages."""
@@ -32,31 +38,82 @@ logging.addLevelName(HINT, "HINT")
 logging.addLevelName(DEPRECATED, "DEPRECATED")
 
 
-class _StderrHandler(logging.StreamHandler):  # type: ignore[type-arg]
-    """A logging handler hard-coded to `sys.stderr`.
+class LoggingHandler(logging.Handler):
+    """coBib's `logging.Handler`."""
 
-    The reason for explicitly deriving this class, is that Python's `logging.StreamHandler` does not
-    respect stream redirection. However, for coBib's TUI this is an important requirement which can
-    be achieved by a runtime check of the stream during the `emit` method.
-    """
+    FORMAT: str = "[%(levelname)s] %(message)s"
+    """The Formatter `fmt` string."""
+
+    DATE_FORMAT: str = "%H:%M:%S"
+    """The Formatter `datefmt` string."""
+
+    def __init__(self, ui: UI, level: int = logging.INFO) -> None:
+        """Initializes the handler.
+
+        Args:
+            ui: the running UI instance.
+            level: the default logging level to be displayed.
+        """
+        super().__init__(level=level)
+
+        self.ui = ui
+
+        formatter = logging.Formatter(fmt=self.FORMAT, datefmt=self.DATE_FORMAT)
+        self.setFormatter(formatter)
+
+        for handler in self.ui.root_logger.handlers[:]:
+            if isinstance(handler, logging.StreamHandler) or (
+                # NOTE: the second condition is required to ensure the unittests pass in Python 3.9
+                isinstance(handler, LoggingHandler) and not isinstance(handler, self.__class__)
+            ):
+                self.ui.root_logger.removeHandler(handler)
+                handler.close()
+
+        self.ui.root_logger.addHandler(self)
+
+    @override
+    def format(self, record: logging.LogRecord) -> Text:  # type: ignore[override]
+        message = super().format(record)
+
+        style = ""
+        if record.levelno >= logging.CRITICAL:
+            style = "bold red"
+        elif record.levelno >= DEPRECATED:
+            style = "bold yellow"
+        elif record.levelno >= logging.ERROR:
+            style = "red"
+        elif record.levelno >= HINT:
+            style = "bold green"
+        elif record.levelno >= logging.WARNING:
+            style = "yellow"
+        elif record.levelno >= logging.INFO:
+            style = "green"
+        elif record.levelno >= logging.DEBUG:  # pragma: no branch
+            style = "blue"
+
+        message = message.replace(
+            f"[{record.levelname}]", f"[{style}][{record.levelname}][/{style}]"
+        )
+
+        text = Text.from_markup(message)
+        return text
 
     @override
     def emit(self, record: logging.LogRecord) -> None:
-        self.stream = sys.stderr
-        super().emit(record)
+        PromptConsole.get_instance().log(self.format(record))
 
 
 def get_stream_handler(
     level: int = logging.WARNING,
 ) -> logging.StreamHandler:  # type: ignore[type-arg]
-    """Returns a basic StreamHandler logging to `sys.stderr`.
+    """Returns a basic StreamHandler logging to a `StringIO` stream.
 
     Args:
         level: the logging level of this handler.
     """
     formatter = logging.Formatter(fmt="[%(levelname)s] %(message)s")
 
-    handler = _StderrHandler()
+    handler = logging.StreamHandler(stream=StringIO())
     handler.setLevel(level)
     handler.setFormatter(formatter)